Preliminary refactor.
