+++ libstainless_macros/Cargo.toml
@@ -8,7 +8,7 @@ categories = ["development-tools"]
 
 [lib]
 name = "stainless_macros"
-path = "lib.rs"
+path = "src/lib.rs"
 test = false
 doctest = false
 proc-macro = true
